- [x] Move duplicate voronoi function (ca2tel_voronoi_coverage, ca2tel_voronoi_age) into unique place (ca2tel)
